Rename variables

+++ src/main/java/com/ordersmicroservice/orders_microservice/dto/CartProductDto.java
@@ -6,16 +6,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class CartProductDto {
-
     private Long id;
     private String productName;
     private Integer quantity;
-    private Double price;
-
-
+    private BigDecimal price;
 }
